+++ src/auth/oauth.ts
@@ -1,5 +1,5 @@
 import { Ed25519Keypair } from '@mysten/sui/keypairs/ed25519';
-import EnokiClient from './enoki';
+import EnokiClient, { ENOKI_PUBLIC_API_KEY } from './enoki';
 import { AuthProvider, OAuthConfig } from '../types/auth';
 import { BadRequest } from '../errors/bad-request';
 import { logger } from '../logger';
@@ -7,19 +7,24 @@ import AkordApi from '../api/akord-api';
 import { Unauthorized } from '../errors/unauthorized';
 import { DEFAULT_STORAGE, JWTClient } from './jwt';
 import { Env } from '../env';
+import { retry } from '../api/api-client';
+import { throwError } from '../errors/error-factory';
 
 export const AuthProviderConfig = {
   "Google": {
     "CLIENT_ID": "52977920067-5hf88uveake073ent9s5snn0d8kfrf0t.apps.googleusercontent.com",
     "AUTH_URL": "https://accounts.google.com/o/oauth2/v2/auth",
+    "SCOPES": "openid email profile",
   },
   "Facebook": {
     "CLIENT_ID": "1030800888399080",
     "AUTH_URL": "https://www.facebook.com/v11.0/dialog/oauth",
+    "SCOPES": "email public_profile",
   },
   "Twitch": {
     "CLIENT_ID": "g924m6acqg8w9gw9hxstdl5q2uugup",
     "AUTH_URL": "https://id.twitch.tv/oauth2/authorize",
+    "SCOPES": "openid user:read:email",
   },
   // TODO: configure Apple
   // "Apple": {
@@ -36,6 +41,7 @@ class OAuth {
   private authProvider: AuthProvider;
 
   private jwtClient: JWTClient;
+  private enokiClient: EnokiClient;
 
   private storage: Storage;
 
@@ -46,15 +52,13 @@ class OAuth {
     if (!AuthProviderConfig[config.authProvider]) {
       throw new BadRequest("Unsupported authProvider, valid providers: " + Object.keys(AuthProviderConfig).join(', '));
     }
-    if (!config.redirectUri) {
-      throw new BadRequest("Missing redirect uri, please provide your app auth callback URL.");
-    }
     this.env = config.env;
     this.redirectUri = config.redirectUri;
     this.authProvider = config.authProvider;
     this.clientId = config.clientId || AuthProviderConfig[this.authProvider].CLIENT_ID;
     this.storage = config.storage || DEFAULT_STORAGE;
     this.jwtClient = new JWTClient({ storage: this.storage, env: this.env });
+    this.enokiClient = new EnokiClient({ apiKey: ENOKI_PUBLIC_API_KEY });
   }
 
   // redirect to OAuth provider's authorization URL
@@ -74,17 +78,25 @@ class OAuth {
       // exchange the authorization code for tokens
       await this.exchangeCodeForTokens(code);
 
-      const enokiClient = new EnokiClient({ apiKey: "enoki_public_ab093e91616fc8fe6125017946ea2548" });
-
-      const { address } = await enokiClient.getZkLogin(this.jwtClient.getIdToken());
+      const { address } = await retry(async () => {
+        try {
+          return this.enokiClient.getZkLogin(this.jwtClient.getIdToken());
+        } catch (error) {
+          throwError(error.response?.status, error.response?.data?.msg, error);
+        }
+      });
+      this.jwtClient.setAddress(address);
       return { address };
     } else {
-      throw new Error('Authorization code not found.');
+      throw new Unauthorized("Authorization code not found.");
     }
   }
 
   async exchangeCodeForTokens(code: string) {
     try {
+      if (!this.redirectUri) {
+        throw new BadRequest("Missing redirect uri, please provide your app auth callback URL.");
+      }
       const { idToken, accessToken, refreshToken } = await new AkordApi({ debug: true, logToFile: true, env: this.env }).generateJWT({
         authProvider: this.authProvider,
         grantType: "code",
@@ -104,16 +116,23 @@ class OAuth {
     // generate ephemeral key pair
     const ephemeralKeyPair = new Ed25519Keypair();
 
-    const enokiClient = new EnokiClient({ apiKey: "enoki_public_ab093e91616fc8fe6125017946ea2548" });
+    const createZkLoginResponse = await retry(async () => {
+      try {
+        return this.enokiClient.createZkLoginNonce(ephemeralKeyPair);
+      } catch (error) {
+        throwError(error.response?.status, error.response?.data?.msg, error);
+      }
+    });
 
-    const createZkLoginResponse = await enokiClient.createZkLoginNonce(ephemeralKeyPair);
-
+    if (!this.redirectUri) {
+      throw new BadRequest("Missing redirect uri, please provide your app auth callback URL.");
+    }
     const params = new URLSearchParams({
       nonce: createZkLoginResponse.nonce,
       client_id: AuthProviderConfig[this.authProvider].CLIENT_ID,
       redirect_uri: this.redirectUri,
       response_type: "code",
-      scope: "openid",
+      scope: AuthProviderConfig[this.authProvider].SCOPES,
       access_type: "offline",
       prompt: "consent"
     });
@@ -126,7 +145,7 @@ class OAuth {
     try {
       const refreshToken = this.jwtClient.getRefreshToken();
       if (!refreshToken) {
-        throw new Unauthorized("Session expired. Please login again.");
+        throw new Unauthorized("Session expired. Please log in again.");
       }
 
       const { idToken, accessToken } = await new AkordApi({ debug: true, logToFile: true }).generateJWT({
@@ -139,7 +158,8 @@ class OAuth {
       this.jwtClient.setIdToken(idToken);
 
     } catch (error) {
-      throw new Error(`Failed to refresh tokens: ${error}`);
+      Logger.error(error);
+      throw new Unauthorized("Failed to refresh tokens.");
     }
   };
 }